Fix Logic Pro buffer size bug.
Logic Pro has an annoying quirk where the buffer size reported by the ho
Logic Pro has an annoying quirk where the buffer size reported by the
host to the plug-in is *always* 1024, regardless of the actually chosen
buffer size. The only way to guarantee a correct value is by checking
the actual number of samples inside the buffer we are given.